requirements.txt
# PDF parsing
PyMuPDF>=1.24.0

# LLM providers
openai>=1.30.0

# Data validation
pydantic>=2.0.0

# Knowledge graph
rdflib>=7.0.0

# CLI & display
rich>=13.0.0
tabulate>=0.9.0

# Environment variables
python-dotenv>=1.0.0
